Create default blocks in both AddPage and CreateSite (homepage) actions.
